Join.Interpretation.Simple.Types

Documentation

data ChanId

Unique identifier of a Channel.

type ReplyId = Maybe ChanId

Potential identifier on which a Reply should be sent.

data MsgData

Encoding of a value send on a Channel.

data Msg

A single encoded value as sent on a Channel, tagged with a possible return ChanId.

type Msgs = [Msg]

Many Msg's.

data MsgQueue

Map ChannelId's to all the Msg's waiting on them.

newChanId :: IO ChanId

Get a new, ChanId, guaranteed to be unique.

mkChanId :: Channel t a -> ChanId

getChanId :: ChanId -> Int

Get an Int representation of a ChanId.

mkMsgData :: Serialize a => a -> MsgData

unMsgData :: MsgData -> ByteString

unsafeDecodeMsgData :: Serialize a => MsgData -> a

mkMsg :: MsgData -> ReplyId -> Msg

unMsg :: Msg -> (MsgData, ReplyId)

mkMsgQueue :: MsgQueue

data ChanMatch

Encodes a single item on the LHS of a Def Pattern. Match the first Msg sent to a ChanId/ first Msg equal to MsgData.

type ChanMatches = [ChanMatch]

Encodes the LHS of a Def Pattern. Match when all individual ChanMatch's simultaneously match.

data MatchRule

Encodes an entire Def Pattern. When the ChanMatches match, a function should be triggered.

Constructors

forall f . Apply f => MatchRule ChanMatches f

type MatchRules = [MatchRule]

mkChanMatch :: SubPattern t p => t -> ChanMatch

Lift a Language SubPattern into an internal ChanMatch.

mkChanMatches :: Pattern t p => t -> ChanMatches

Lift a Language Pattern into an internal ChanMatches.

mkMatchRule :: (Apply f, Pattern t f) => t -> f -> MatchRule

Lift the LHS and RHS of a Language Def into an internal MatchRule.

registerChanId :: ChanId -> MsgQueue -> MsgQueue

Register a new ChanId (With no waiting Msg's) in a MsgQueue.

registerMsg :: ChanId -> Msg -> MsgQueue -> MsgQueue

Register, on a ChanId, a new Msg in a MsgQueue.

registerMatchRule :: Apply f => ChanMatches -> f -> MatchRules -> MatchRules

Given ChanMatches and a corresponding trigger function, register it in a MatchRules collection.

tryMsgMatch :: Maybe MsgData -> Msgs -> Maybe (Msgs, Msg)

With a possible MsgData to require equality with, extract the first matching Msg from some Msgs.

tryChanMatch :: ChanMatch -> MsgQueue -> Maybe (MsgQueue, Msg)

Given a ChanMatch, attempt to extract a matching Msg within a MsgQueue.

tryChanMatches :: ChanMatches -> MsgQueue -> Maybe (MsgQueue, Msgs)

Given a collection of ChanMatch's, extract a set of matching Msgs when each ChanMatch is successful.

tryMatchRule :: MatchRule -> MsgQueue -> Maybe (MsgQueue, Msgs)

Given a MatchRule and MessageQueues, test the rule, if successful returning the updated MessageQueues and the matching Messages.

extractReplyIds :: ChanMatches -> Msgs -> [(ChanId, ChanId)]