Join.Language.Types.Channel

Description

This module defines a Channel type, a representation of typed Join-Calculus Channels.

Channels themselves do not provide a means of queuing messages, what they provide is a means for a Join-Calculus program to:

Track where it should route messages sent on the channel.

Enforce the type of messages allowed.

Determine whether reply-values are allowed/ expected.

The first point is facilitated by the id each Channel is constructed with. The second and third, by the type variables.

inferSync allows an implementation to infer whether a newly created Channel should be Asynchronous or Synchronous based upon it's usage. For example, a reply function could be defined to operate exclusively on Synchronous Channels, then, if reply is called on a channel, it can be inferred that it is intended to be Synchronous and the user can omit type signatures. By constraining the synchronicity type variable in an implementation sufficiently, type annotations may be able to be avoided entirely.

Documentation

data Channel s a

A Channel uniquely identifies a port of communication.

The type parameter a denotes the type of values accepted.

The type parameter s is of kind Synchronicity and denotes whether values should be sent Asynchronously (a regular channel as defined by the core calculus) or Synchronously, where a return value is expected.

Channels are constructed with an Int parameter which serves as it's unique ID. Interpreters should ensure these are unique.

data Synchronicity

Synchronicity tag Type & Kind.

Constructors

A
Asynchronous.

S
Synchronous.

getId :: Channel s a -> Int

Extract the Id of a Channel.

class InferSync s where

Closed Class of Synchronicity-kinded types which may be used to infer how to construct a corresponding Channel.

I.E. When:

  i :: Int
  c = inferSync i :: Channel s a

and in a context where s is constrained to A or S, then inferSync will create the corresponding type of Channel.

Methods

inferSync :: Serialize a => Int -> Channel s a