Join.Language.Types.SyncVal

Description

This module defines a SyncVal type, which can be used to encapsulate a value which may not have been computed yet.

The type may be used to wrap the return value of synchronous calls. The advantage of doing so would be that other actions could be performed after the synchronous call and before the point where the value is required. This could eliminate unnecessary waiting.

Documentation

data SyncVal a

Represents a Synchronous value which may or may not exist yet.

new :: IO (SyncVal a)

Create a new empty SyncVal.

wait :: SyncVal a -> a

Synchronously wait for a value on a SyncVal.

write :: SyncVal a -> a -> IO ()

Write a value to a SyncVal. Throws Data.IVar.Simple.BlockedIndefinitelyOnIVar exception if a value has already been written.